+++ tests/test_sensor_A5_04_01.py
@@ -13,7 +13,7 @@ Entity.schedule_update_ha_state = mock.Mock(return_value=None)
 # EltakoBinarySensor.hass.bus.fire is mocked by class HassMock
 
 
-class TestSensor_A5_04_02(unittest.TestCase):
+class TestSensor_A5_04_01(unittest.TestCase):
 
     msg1 = Regular4BSMessage (address=b'\xFF\xFF\x00\x80', data=b'\xaa\x00\x00\x0d', status=0x00)
     
@@ -21,7 +21,7 @@ class TestSensor_A5_04_02(unittest.TestCase):
         gateway = GatewayMock()
         dev_id = AddressExpression.parse("FF-FF-00-80")
         dev_name = "device name"
-        dev_eep = EEP.find("A5-04-02")
+        dev_eep = EEP.find("A5-04-01")
         s = EltakoTemperatureSensor(Platform.SENSOR, gateway, dev_id, dev_name, dev_eep)
         return s
     
@@ -29,7 +29,7 @@ class TestSensor_A5_04_02(unittest.TestCase):
         gateway = GatewayMock()
         dev_id = AddressExpression.parse("FF-FF-00-80")
         dev_name = "device name"
-        dev_eep = EEP.find("A5-04-02")
+        dev_eep = EEP.find("A5-04-01")
         s = EltakoHumiditySensor(Platform.SENSOR, gateway, dev_id, dev_name, dev_eep)
         return s
     
@@ -37,7 +37,7 @@ class TestSensor_A5_04_02(unittest.TestCase):
         s_temp = self.create_temperature_sensor()
 
         s_temp.value_changed(self.msg1)
-        self.assertEqual(s_temp.native_value, -20.0)
+        self.assertEqual(s_temp.native_value, 0.0)
 
     def test_humidity_sensor_A5_04_02(self):
         s_hum = self.create_humidity_sensor()