+++ changes.md
@@ -1,7 +1,7 @@
 # Changes and Feature List
 
 ## Version 1.4.1 Support for sending arbritrary messages
-* Added Service for sending arbritrary ESP2 messages
+* Added Service for sending arbritrary EnOcean (ESP2) messages. Intended to be used in conjunction with [Home Assistant Automations](https://www.home-assistant.io/getting-started/automation/).
 * 🐞 Fix for TargetTemperatureSensor (EEP: A5-10-06 and A5-10-12)
 * 🐞 Fix for unknow cover positions and intermediate state + unit-tests added.
 * Unit-Tests added and improved for EEP A5-04-01, A5-04-02, A5-10-06, A5-10-12, A5-13-01, and F6-10-00.